requirements.txt
# The requirements listed here are for development.
# Requirements of your napari plugin should be listed in setup.cfg
# See also: https://caremad.io/posts/2013/07/setup-vs-requirement/
ipywidgets
sphinx
nbsphinx<0.8.8
sphinx-jsonschema<1.19.0
Jinja2<3.1.0
magicgui
pytest-qt
platformdirs
taskipy
-e .
